% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{AUC}
\alias{AUC}
\title{Compute the AUC for ROC values}
\usage{
AUC(x, ...)
}
\arguments{
\item{x}{numeric, vector of \code{TPR} or \code{ROC} object}

\item{...}{other arguments for \code{\link[AUC]{auc}}}
}
\value{
numeric Area Under Curve as per \code{\link[AUC]{auc}}
}
\description{
We use the \href{https://en.wikipedia.org/wiki/Trapezoidal_rule}{trapezoidal rule}
}
