% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{FPR}
\alias{FPR}
\title{False Positive Rate \code{FP/N} aka "specificity"}
\usage{
FPR(x, ...)
}
\arguments{
\item{x}{tibble with columns label (0/1) and pred (0-1)}

\item{...}{further arguments for \code{\link[AUC]{sensitivity}}}
}
\value{
a list as per \code{\link[AUC]{sensitivity}}
}
\description{
See more \href{https://en.wikipedia.org/wiki/Sensitivity_and_specificity}{here}
}
