% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{gather_plots}
\alias{gather_plots}
\title{Function to gather and patchwork the plots}
\usage{
gather_plots(df)
}
\arguments{
\item{df}{list of dataframes to apply the geom_response to}
}
\value{
a patchworked list of ggplots
}
\description{
Function to gather and patchwork the plots
}
