% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{geom_likelihood}
\alias{geom_likelihood}
\title{Function to plot based on value type}
\usage{
geom_likelihood(df)
}
\arguments{
\item{df}{list of df's}
}
\value{
list of ggplot objects
}
\description{
Function to plot based on value type
}
