% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pauc.R
\name{pAUC}
\alias{pAUC}
\title{Compute AUC for predicted presences}
\usage{
pAUC(x, y, ...)
}
\arguments{
\item{x}{object containing all predicted ranked suitability values, such as
values produced by the \code{predict()} function. This may be a vector or \code{stars} object.}

\item{y}{object containing predicted ranked suitability values for presence points only.
This may be a vector or, if \code{x} is a \code{stars} object it muct be a \code{sf} \code{POINT} object.}

\item{...}{other arguments for \code{auc_raster} or \code{auc_vector}.}
}
\value{
a list of
\itemize{
\item{fpa, a vector of fractional predicted area}
\item{sensitivity,a vector of sensitivity (1-omission rate)}
\item{area, scalar AUC value}
\item{fn, count f non-missing values}
\item{vn, count v non-missing values}
}
}
\description{
Compute AUC for predicted presences
}
