% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pauc.R
\name{pauc_raster}
\alias{pauc_raster}
\title{Given a stars object and presence point locations compute the pAUC}
\usage{
pauc_raster(
  x,
  y,
  time_column = attr(y, "time_column") \%||\% attr(y, "time_col"),
  ...
)
}
\arguments{
\item{x}{a stars object}

\item{y}{sf point object}

\item{time_column}{char or num, indicates the time-based column in \code{y} that matches the
time dimension in \code{x}}

\item{...}{further arguments for \code{auc_vector}}
}
\value{
a list of
\itemize{
\item{fpa, a vector of fractional predicted area}
\item{sensitive,a vector of sensitivity (1-omission rate)}
\item{area, scalar AUC value}
}
}
\description{
Given a stars object and presence point locations compute the pAUC
}
\seealso{
\code{auc_vector}
}
