% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot.ROC}
\alias{plot.ROC}
\title{Plot ROC curve}
\usage{
\method{plot}{ROC}(
  x,
  xlab = "False Positive Rate (Specificity)",
  ylab = "True Positive Rate (Sensitivity)",
  title = "ROC",
  use = "ggplot",
  ...
)
}
\arguments{
\item{x}{RC object}

\item{xlab}{char, label for x axis}

\item{ylab}{char, label for y axis}

\item{title}{char title for the plot}

\item{use}{char, one of 'base' or 'ggplot' (default) or 'ggplot2'}

\item{...}{ignored}
}
\value{
ggplot2 object if \code{use} is
 'ggplot2' or 'ggplot'
}
\description{
Plot ROC curve
}
