% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_maxnet}
\alias{read_maxnet}
\title{Read a maxnet model file}
\usage{
read_maxnet(filename, ...)
}
\arguments{
\item{filename}{char, filename of the file to read}

\item{...}{other arguments for \code{\link[base]{readRDS}}}
}
\value{
maxnet object
}
\description{
Read a maxnet model file
}
