% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{ROC}
\alias{ROC}
\title{Compute Receiver Operator Curve (ROC) space values}
\usage{
ROC(x)
}
\arguments{
\item{x}{tibble with columns label (0/1) and pred (0-1)}
}
\value{
list as per \code{\link[AUC]{roc}} with 'area' added
}
\description{
Compute Receiver Operator Curve (ROC) space values
}
