% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{TPR}
\alias{TPR}
\title{Compute True Positive Rate \code{TP/P} aka "sensitivity"}
\usage{
TPR(x, ...)
}
\arguments{
\item{x}{tibble with columns label (0/1) and pred (0-1)}

\item{...}{further arguments for \code{\link[AUC]{sensitivity}}}
}
\value{
list as per \code{\link[AUC]{sensitivity}}
}
\description{
See more \href{https://en.wikipedia.org/wiki/Sensitivity_and_specificity}{here}
}
