% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{variable_importance}
\alias{variable_importance}
\title{Compute variable importance}
\usage{
variable_importance(
  x,
  y,
  n = 5,
  type = c("link", "exponential", "cloglog", "logistic")[3]
)
}
\arguments{
\item{x}{maxnet model}

\item{y}{table of occurence and background environmental data (data.frame,
tibble, or matrix with column names)}

\item{n}{num, number of interations}

\item{type}{char, the model type - see \code{\link[maxnet]{maxent}}}
}
\description{
This is adapted from Peter D Wilson's \href{https://github.com/peterbat1/fitMaxnet}{fitMaxnet R package}
}
