% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{variable_names}
\alias{variable_names}
\title{Retrieve variable names}
\usage{
variable_names(x)
}
\arguments{
\item{x}{maxnet model object}
}
\value{
character vector of one or more names
}
\description{
Retrieve variable names
}
