% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_maxnet}
\alias{write_maxnet}
\title{Save a maxnet model as a serialize R data object}
\usage{
write_maxnet(x, filename, ...)
}
\arguments{
\item{x}{maxnet model}

\item{filename}{character, the name of the file to save (use .rds extension recommended)}

\item{...}{other arguments for \code{\link[base]{saveRDS}}}
}
\value{
the input x invisibly
}
\description{
This provides a pipeable workflow.
}
