% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{AUC}
\alias{AUC}
\title{Compute the AUC for ROC values}
\usage{
AUC(x, ...)
}
\arguments{
\item{x}{tibble with columns label (0/1) and pred (0-1)}

\item{...}{other arguments for \code{\link[AUC]{auc}}}
}
\value{
numeric Area Under Curve as per \code{\link[AUC]{auc}}
}
\description{
Compute the AUC for ROC values
}
