% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxnetic-package.R
\docType{package}
\name{maxnetic}
\alias{maxnetic-package}
\alias{maxnetic}
\title{maxnetic package}
\description{
Supplementary tools extending the maxnet package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bigelowlab/maxnetic}
}

}
